Fix Terraform state lock not getting released when CI would get canceled
So just don't use a `concurrency` statement to cancel redundant CI runs.

Would often end up in Terraform not releasing the lock that that it
acquired for running the plan, and we would need to manually delete the
lock from the DynamoDB table.

Based on GHA docs it should handle it gracefully but it simply does not.

Quoting my comment:
> I can also confirm that using `cancel-in-progress: true` with a GitHub
> Actions CI pipeline that runs `terraform plan` does not work well. If
> the plan step is canceled it will just immediately terminate and not
> gracefully release the lock.
>
> I've tested locally sending `SIGINT` and `SIGTERM` signals to a
> running `terraform plan` and it **does** exit gracefully and releases
> the lock. GitHub Actions seems to be sending some other, more
> aggressive signal, when killing the workflow.
>
> https://github.community/t/graceful-job-termination/121103/6